+++ examples/derive.rs
@@ -1,16 +1,17 @@
-use rustsbi::{HartMask, RustSBI};
+use rustsbi::{HartMask, MachineInfo, RustSBI};
 use sbi_spec::binary::SbiRet;
 
 #[derive(RustSBI)]
 struct MySBI {
     fence: MyFence,
+    info: MyMachineInfo,
 }
 
 struct MyFence;
 
 impl rustsbi::Fence for MyFence {
     fn remote_fence_i(&self, _: HartMask) -> SbiRet {
-        println!("remote fence i");
+        println!("remote fence i called");
         SbiRet::success(0)
     }
 
@@ -23,13 +24,28 @@ impl rustsbi::Fence for MyFence {
     }
 }
 
+struct MyMachineInfo;
+
+impl MachineInfo for MyMachineInfo {
+    fn mvendorid(&self) -> usize {
+        0x100
+    }
+
+    fn marchid(&self) -> usize {
+        0x200
+    }
+
+    fn mimpid(&self) -> usize {
+        0x300
+    }
+}
+
 fn main() {
-    let sbi = MySBI { fence: MyFence };
+    let sbi = MySBI {
+        fence: MyFence,
+        info: MyMachineInfo,
+    };
     sbi.handle_ecall(sbi_spec::rfnc::EID_RFNC, 0, [0; 6]);
-    let spec_version = sbi.handle_ecall(
-        sbi_spec::base::EID_BASE,
-        sbi_spec::base::GET_SBI_SPEC_VERSION,
-        [0; 6],
-    );
-    println!("spec version: {:x?}", spec_version.value);
+    let sbi_impl_id = sbi.handle_ecall(0x10, 0x1, [0; 6]);
+    println!("SBI implementation ID: {:x?}", sbi_impl_id.value);
 }